requirements.txt
scikit-image
tensorflow
numpy
pillow